LayoutTests/platform/gtk/accessibility/aria-option-role.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="listbox">
<div role="option">option 1</div>
<div role="option" aria-label="label 2">option 2</div>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the aria 'option' role works as expected.");

    if (window.accessibilityController) {

          var body = document.getElementById("body");
          body.focus();

          var listBox = accessibilityController.focusedElement.childAtIndex(0);
          var firstChild = listBox.childAtIndex(0);
          var secondChild = listBox.childAtIndex(1);
          shouldBe("firstChild.role", "'AXRole: list item'");
          shouldBe("firstChild.title", "'AXTitle: option 1'");

          shouldBe("secondChild.role", "'AXRole: list item'");
          shouldBe("secondChild.description", "'AXDescription: label 2'");

          shouldBe("firstChild.childrenCount", "0");
          shouldBe("secondChild.childrenCount", "0");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
